# academicplanner-backend/docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "AcademicPlanner API",
    "version": "0.1.0"
  },
  "paths": {
...
    "/user/login": {
      "post": {
        "summary": "Login",
        "operationId": "login_user_login_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LoginRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TokenResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/schedule": {
      "post": {
        "summary": "Schedule",
        "operationId": "schedule_schedule_post",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ScheduleResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/user/refresh-token": {
      "post": {
        "summary": "Refresh",
        "operationId": "refresh_user_refresh_token_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RefreshRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TokenResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "AcademicClass": {
        "properties": {
          "id": {
            "type": "string",
            "minLength": 1,
            "title": "Id"
          },
          "subjectCode": {
            "type": "string",
            "minLength": 1,
            "title": "Subjectcode"
          },
          "subjectName": {
            "type": "string",
            "minLength": 1,
            "title": "Subjectname"
          },
          "section": {
            "type": "string",
            "title": "Section"
          },
          "professor": {
            "type": "string",
            "title": "Professor",
            "default": ""
          },
          "day": {
            "type": "integer",
            "maximum": 7.0,
            "minimum": 1.0,
            "title": "Day"
          },
          "startTime": {
            "type": "string",
            "pattern": "^([01]\\d|2[0-3]):[0-5]\\d$",
            "title": "Starttime"
          },
          "endTime": {
            "type": "string",
            "pattern": "^([01]\\d|2[0-3]):[0-5]\\d$",
            "title": "Endtime"
          },
          "location": {
            "type": "string",
            "title": "Location",
            "default": ""
          }
        },
        "type": "object",
        "required": [
          "id",
          "subjectCode",
          "subjectName",
          "section",
          "day",
          "startTime",
          "endTime"
        ],
        "title": "AcademicClass"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "LoginRequest": {
        "properties": {
          "studentId": {
            "type": "string",
            "maxLength": 100,
            "minLength": 1,
            "title": "Studentid"
          },
          "password": {
            "type": "string",
            "maxLength": 256,
            "minLength": 1,
            "format": "password",
            "title": "Password",
            "writeOnly": true
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "studentId",
          "password"
        ],
        "title": "LoginRequest"
      },
      "RefreshRequest": {
        "properties": {
          "refreshToken": {
            "type": "string",
            "maxLength": 512,
            "minLength": 1,
            "format": "password",
            "title": "Refreshtoken",
            "writeOnly": true
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "refreshToken"
        ],
        "title": "RefreshRequest"
      },
      "ScheduleResponse": {
        "properties": {
          "student": {
            "$ref": "#/components/schemas/Student"
          },
          "fetchedAt": {
            "type": "string",
            "format": "date-time",
            "title": "Fetchedat"
          },
          "classes": {
            "items": {
              "$ref": "#/components/schemas/AcademicClass"
            },
            "type": "array",
            "title": "Classes"
          },
          "unscheduledSubjects": {
            "items": {
              "$ref": "#/components/schemas/UnscheduledSubject"
            },
            "type": "array",
            "title": "Unscheduledsubjects"
          }
        },
        "type": "object",
        "required": [
          "student",
          "fetchedAt",
          "classes"
        ],
        "title": "ScheduleResponse"
      },
      "Student": {
        "properties": {
          "id": {
            "type": "string",
            "maxLength": 64,
            "minLength": 1,
            "title": "Id"
          },
          "isPino": {
            "type": "boolean",
            "title": "Ispino",
            "default": false
          }
        },
        "type": "object",
        "required": [
          "id"
        ],
        "title": "Student"
      },
      "TokenResponse": {
        "properties": {
          "accessToken": {
            "type": "string",
            "title": "Accesstoken"
          },
          "refreshToken": {
            "type": "string",
            "title": "Refreshtoken"
          },
          "tokenType": {
            "type": "string",
            "title": "Tokentype",
            "default": "bearer"
          },
          "expiresIn": {
            "type": "integer",
            "title": "Expiresin",
            "default": 900
          }
        },
        "type": "object",
        "required": [
          "accessToken",
          "refreshToken"
        ],
        "title": "TokenResponse"
      },
      "UnscheduledSubject": {
        "properties": {
          "reason": {
            "type": "string",
            "enum": [
              "not_reported",
              "to_be_announced",
              "asynchronous"
            ],
            "title": "Reason",
            "default": "not_reported"
          },
          "id": {
            "type": "string",
            "minLength": 1,
            "title": "Id"
          },
          "subjectCode": {
            "type": "string",
            "minLength": 1,
            "title": "Subjectcode"
          },
          "subjectName": {
            "type": "string",
            "minLength": 1,
            "title": "Subjectname"
          },
          "section": {
            "type": "string",
            "title": "Section"
          },
          "professor": {
            "type": "string",
            "title": "Professor",
            "default": ""
          },
          "location": {
            "type": "string",
            "title": "Location",
            "default": ""
          }
        },
        "type": "object",
        "required": [
          "id",
          "subjectCode",
          "subjectName",
          "section"
        ],
        "title": "UnscheduledSubject"
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  }
}